Add API key authentication for the MaaS Operators
User/Maas operator is authorized using an API key which is
read from the authorization header.

DRFs existing token authentication functionality is used, which
means linking a Django user to a MaasOperator instance. This is
why the API token generation logic is refactored so that DRF provided
logic can be used instead of rolling our own token authentication logic
to use MaasOperator objects.

IsMaasOperator permission can be used in a view to require that
the user of the API endpoint is a MaaS operator (i.e. user has
a link to a MaasOperator instance).

Refs MAAS-28