Change arrival and departures times to durations
Because we need to support GTFS times that go past midnight
(like 26:00), we cannot use TimeFields for StopTime arrival_time and
departure_time. While they are not actually durations, using
DurationFields for them is easier than implementing a custom field.

As a side effect of this change, agency timezone is now properly taken
into account also.