Return distinct stop times in StopSerializer
The API was not returning distinct StopTimes (departures in the API)
when a stop appeared more than once on a trip. If the same stop appeared
more than once on a trip (when visiting the same place more than once
during a trip), the API didn't properly return the stop times for
additional visits to the same stop.

Refs MAAS-101