Add StopTime.stops_after_this field
Added field `stops_after_this` to `StopTime` model which tells how many
stop times there are on the trip after the stop time in question. The
main purpose of the field is to be able to tell which stop times are
the last ones. The field is populated once after a feed import. This
denormalized field makes queries relating to stop times' last
departures a lot easier and lighter.